+++ rift-engine/rift/ir/symbol_table.txt
